Help users avoid running a second daemon process
Issue https://github.com/pop-os/upgrade/issues/99 describes a
source of confusion where users can run a daemon process that
the release command will never connect to if the `pop-upgrade`
service is already running in the background.

This fix points users to try `journalctl` first or alternatively
that they should stop the service first before running the
daemon in the foreground.